+++ xalpha/backtest.py
@@ -6,12 +6,12 @@ modules for dynamical backtesting framework
 import pandas as pd
 
 from xalpha.info import fundinfo, mfundinfo, cashinfo
-from xalpha.trade import trade
+from xalpha.trade import trade, itrade
 from xalpha.multiple import mul, mulfix
 from xalpha.cons import yesterdayobj, avail_dates
 from xalpha.exceptions import TradeBehaviorError, FundTypeError
 from xalpha.cons import opendate_set, next_onday, convert_date
-from xalpha.universal import vinfo
+from xalpha.universal import vinfo, get_daily
 
 
 class GlobalRegister:
@@ -188,6 +188,9 @@ class BTE:
             remtable = self.trades[code].remtable
             remtable = remtable[remtable["date"] <= self.lastdates[code]]
             self.lastdates[code] = date
+            if is_value is False:
+                tdf = self.infos[code].price
+                value = value * tdf[tdf["date"] == date].iloc[0].netvalue
             df2 = pd.DataFrame([[date, value]], columns=["date", self.get_code(code)])
             df = df.append(df2)
             self.trades[code] = trade(
@@ -438,3 +441,73 @@ class Balance(BTE):
                     )  # 赎回份额考虑赎回费估算为千五，会导致末态并非完全平衡
                 elif delta < 0:
                     self.buy(fund, -delta, date)
+
+
+class Grid(BTE):
+    """
+    网格策略
+    """
+
+    def prepare(self):
+        self.df = get_daily(self.kws.get("code"), start=self.start.strftime("%Y%m%d"))
+        self.infos[self.kws.get("code")] = self.get_info(self.kws.get("code"))
+        self.prices = self.kws.get("prices")
+        self.inamount = self.kws.get("inamount")
+        self.outamount = self.kws.get("outamount")
+        self.pos = 0
+        self.cftable = pd.DataFrame([], columns=["date", "cash", "share"])
+
+    def run(self, date):
+        row = self.df[self.df["date"] == date]
+        high = row["high"].item()
+        low = row["low"].item()
+        for p in self.prices[self.pos + 1 :]:
+            if p > low:
+                p0 = p
+                if p0 > high:
+                    p0 = high
+                if self.verbose is True:
+                    print(
+                        f"buy {self.inamount[self.pos]} of {self.kws.get('code')} at {date} with value {p0}"
+                    )
+                self.cftable = self.cftable.append(
+                    pd.DataFrame(
+                        [
+                            [
+                                date,
+                                -p0 * self.inamount[self.pos],
+                                self.inamount[self.pos],
+                            ]
+                        ],
+                        columns=["date", "cash", "share"],
+                    ),
+                    ignore_index=True,
+                )
+                self.pos += 1
+            else:
+                break
+        for p in self.prices[: self.pos][::-1]:
+            if p < high:
+                if self.verbose is True:
+                    print(
+                        f"sell {self.outamount[self.pos - 1]} of {self.kws.get('code')} at {date} with value {p}"
+                    )
+                self.cftable = self.cftable.append(
+                    pd.DataFrame(
+                        [
+                            [
+                                date,
+                                p * self.outamount[self.pos - 1],
+                                -self.outamount[self.pos - 1],
+                            ]
+                        ],
+                        columns=["date", "cash", "share"],
+                    ),
+                    ignore_index=True,
+                )
+                self.pos -= 1
+            else:
+                break
+        self.trades[self.kws.get("code")] = itrade(
+            self.kws.get("code"), None, self.cftable
+        )